requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic-settings>=2.2.0
redis[asyncio]>=5.0.0
supabase>=2.4.0
httpx>=0.27.0
google-generativeai>=0.8.0
anthropic>=0.40.0
openai>=1.30.0
google-api-python-client>=2.130.0
google-auth>=2.29.0
APScheduler>=3.10.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
asyncpg>=0.29.0
# Vercel
mangum>=0.17.0
